refactor(simulation): extract location lookup feature
