+++ src/test/java/com/renewsim/backend/simulation_service/web/controller/SimulationControllerTest.java
@@ -16,13 +16,10 @@ import com.renewsim.backend.simulation_service.application.detailSimulation.GetR
 import com.renewsim.backend.simulation_service.application.historySimulation.ListUserRealSimulationsUseCase;
 import com.renewsim.backend.simulation_service.application.historySimulation.SimulationHistoryRowResult;
 import com.renewsim.backend.simulation_service.application.historySimulation.UserSimulationListResult;
-import com.renewsim.backend.simulation_service.application.port.out.LocationLookupPort;
 import com.renewsim.backend.simulation_service.application.shared.SimulationDetailsResult;
-import com.renewsim.backend.simulation_service.domain.model.vo.ResolvedLocation;
 import com.renewsim.backend.simulation_service.web.dto.CreateSolarSimulationRequestDTO;
 import com.renewsim.backend.simulation_service.web.dto.ListUserSimulationsResponseDTO;
 import com.renewsim.backend.simulation_service.web.dto.PortfolioDashboardResponseDTO;
-import com.renewsim.backend.simulation_service.web.dto.ResolvedLocationResponseDTO;
 import com.renewsim.backend.simulation_service.web.dto.SimulationDetailsResponseDTO;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -51,8 +48,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @ExtendWith(MockitoExtension.class)
 class SimulationControllerTest {
 
-    @Mock
-    private LocationLookupPort locationLookupPort;
     @Mock
     private CreateRealSimulationUseCase createRealSimulationUseCase;
     @Mock
@@ -157,18 +152,6 @@ class SimulationControllerTest {
 
         verify(listUserRealSimulationsUseCase).getUserSimulations("alice");
         verify(getRealSimulationUseCase, never()).getSimulationById(any(), any(), any(Boolean.class));
-    }
-
-    @Test
-    @DisplayName("reverseGeocode returns resolved location from coordinates")
-    void reverseGeocodeReturnsResolvedLocation() {
-        when(locationLookupPort.resolveLocation(-32.8895, -68.8458))
-                .thenReturn(new ResolvedLocation("Mendoza", "AR", -32.8895, -68.8458));
-
-        ResolvedLocationResponseDTO response = controller.reverseGeocode(-32.8895, -68.8458).getBody();
-
-        assertThat(response).isNotNull();
-        assertThat(response.name()).isEqualTo("Mendoza");
     }
 
     private CreateSolarSimulationRequestDTO request() {
